Extract content fragment logic in file
