+++ front/lib/api/assistant/helpers.ts
@@ -1,0 +1,36 @@
+import type { ConversationType } from "@dust-tt/types";
+import { md5 } from "@dust-tt/types";
+import type { Transaction } from "sequelize";
+
+import { frontSequelize } from "@app/lib/resources/storage";
+import logger from "@app/logger/logger";
+
+/**
+ * To avoid deadlocks when using Postgresql advisory locks, please make sure to not issue any other
+ * SQL query outside of the transaction `t` that is holding the lock.
+ * Otherwise, the other query will be competing for a connection in the database connection pool,
+ * resulting in a potential deadlock when the pool is fully occupied.
+ */
+export async function getConversationRankVersionLock(
+  conversation: ConversationType,
+  t: Transaction
+) {
+  const now = new Date();
+  // Get a lock using the unique lock key (number withing postgresql BigInt range).
+  const hash = md5(`conversation_message_rank_version_${conversation.id}`);
+  const lockKey = parseInt(hash, 16) % 9999999999;
+  await frontSequelize.query("SELECT pg_advisory_xact_lock(:key)", {
+    transaction: t,
+    replacements: { key: lockKey },
+  });
+
+  logger.info(
+    {
+      workspaceId: conversation.owner.sId,
+      conversationId: conversation.sId,
+      duration: new Date().getTime() - now.getTime(),
+      lockKey,
+    },
+    "[ASSISTANT_TRACE] Advisory lock acquired"
+  );
+}
