+++ front/pages/api/v1/w/[wId]/assistant/conversations/index.ts
@@ -12,11 +12,11 @@ import { isLeft } from "fp-ts/lib/Either";
 import * as reporter from "io-ts-reporters";
 import type { NextApiRequest, NextApiResponse } from "next";
 
+import { postNewContentFragment } from "@app/lib/api/assistant/content_fragment";
 import {
   createConversation,
   getConversation,
   normalizeContentFragmentType,
-  postNewContentFragment,
 } from "@app/lib/api/assistant/conversation";
 import { postUserMessageWithPubSub } from "@app/lib/api/assistant/pubsub";
 import { Authenticator, getAPIKey } from "@app/lib/auth";
